Add in conta R files, scripts and associated tests
Summary:
This is the second part of conta check in including main packages. I
can give a personal into on how the code is organized. Hope to put
this code in less than a week to enable conta module in WGS pipeline
before I leave for vacation.

Test Plan:
Easiest way to test is patch the diff, open package from RStudio and
use the interface to build and test.

Reviewers: ablocker, dfilippova, cmelton, jlamping, cbarbacioru

Reviewed By: ablocker, jlamping

Subscribers: razvanm, tkim

Differential Revision: https://phabricator.grailbio.com/D4598

fbshipit-source-id: ca50422