Conta fix error rate on SNP alleles
Summary:
Conta error calculation masks SNPs, and this was causing
errors on SNP alleles to be counted less than expected while
denominator being fully counted. As a fix, for SNP positions
(not the non-SNP positions if conta is run with supplementary
positions), the denominator is reduced by two thirds.

Test Plan:
- Unit tests pass.
- Rerun whole tissue set, error rates and calls are as expected.

Reviewers: rvijayasatya, jtoung

Reviewed By: jtoung

Differential Revision: https://phabricator.grailbio.com/D9849

fbshipit-source-id: 75c09a3