analysis/conta strand specific mode
Summary:
Bisulfite conversion converts most C alleles in the genome
(unless it is methylated) to Ts. Thus, SNPs with a C allele
cannot be used by conta and we were ignoring these SNPs
completely, until now. For SNPs that contain C alleles,
fragments coming from the opposite strand are Gs, and are
not affected by bisulfite conversion. Thus, we can selectively
use these SNPs. Vice versa it applies for SNPs with
G alleles: forward strand is usable, but the reverse strand
is not. On the implementation side, one simple way to achieve
this is to use forward strand and reverse strand counts selectively
on SNPs that are usable for that strand, and then add up the
counts within conta. This requires two count files as input, and
two dbSNP vcf files for filtration. Rest of conta logic is unchanged.

Test Plan: Unit tests pass. Seperate spamtool test on conta with new mode worked.

Reviewers: jtoung

Reviewed By: jtoung

Subscribers: joshnewman, cchang

Differential Revision: https://phabricator.grailbio.com/D16833

fbshipit-source-id: ec058b3