third-party/R: update roxygen2
Summary:
Newer version deals with reference classes better.

But newer version also loads the package before generating
documentation. Which means that we no longer need to use devtools to
generate documentation. The downside of this is that the roclets_deps
attribute of all the packages will now need to change to include the
direct dependencies of the package so it can be loaded by roxygen.

Also resolves some TODOs around r_pkg_test not running on macOS.

And updates DESCRIPTION file to use the new version in RoxygenNote, and
ensures that the Encoding field is present such that Roxygen does not
complain.

And resets timeout values for r_pkg_test and r_unit_test. Most of these
were copy-pasted from other packages and do not need to be moderate.
People can increase the timeouts if their tests really do start failing.

Reviewers: gmeixiong

Reviewed By: gmeixiong

Maniphest Tasks: T12333

Differential Revision: https://phabricator.grailbio.com/D19918

fbshipit-source-id: 8081a7d