analysis/conta: add chrX read and gender call
Summary:
Gender/sex calling requires the use of chrY and chrX read and SNP
counts. Certain thresholds or models need to be determined and abnormal
sex chromosomes needs to be considered. Conta has so far used
only chrY counts when implementing gender logic, but in this revision
we are adding chrX counts, as well as base logic to call gender.

Initial thresholds are trained on a small number of Buckler
samples, but more samples and possibly more complicated logic
(logistic regression) will be needed to improve the calls.
We will also need to detect and report sexual aneuploidies
such as Turner Syndrome (X) and Triple X (XXX), as well as
consider deteriation of sex chromosomes with age.

Test Plan: Unit tests passed.

Reviewers: rcalef

Reviewed By: rcalef

Subscribers: ovenn, ashanmugam, jtoung, cyakym

Differential Revision: https://phabricator.grailbio.com/D21730

fbshipit-source-id: 9f1825f