reflow/methylation: improve caching of conta results
Summary:
In order to ensure that conta can use the cache in the case where the sample ID is different but the data is the same, the conta exec now uses a generic mock sample ID. After conta runs the output files are renamed to have the sample ID in them again. The one side effect here is that the output files that have a sample ID column (or png with sample ID in the title), will have the mock sample ID.

The BUILD file also needed to be changed to run reflow with `-eval=bottomup` and with `-sched=false`. The `-sched=false` is needed because of a bug in reflow where the caching is not done correctly when in `-sched=true` mode.

Ref T23463

Test Plan:
Run determinism test (`determinism_test.sh`) and run tests where there are 3 runs of the same data:

# Run data with cache in write mode.
# Run data with cache on (default state is used), using the same Sample ID and same fastq data.
# Run data with cache on (default state is used), using a different Sample ID and same fastq data.

The cache usage was looked at and compared between these to ensure this fix addressed the major cache "miss".

The attached pdf report from the `determinism_test.sh` shows that the file naming is correct and the attached `.conta.tsv` shows that the sample name is correct. The `.conta.tsv` file is from determinism test `a` and has the date added to the name.

**PDF** {F878501}

**.conta.tsv** {F878510}

Reviewers: sbagaria

Reviewed By: sbagaria

Subscribers: osakarya, rhuang, escott, pgopal, smahadevan

Maniphest Tasks: T23463

Differential Revision: https://phabricator.grailbio.com/D32775

fbshipit-source-id: 0286073