analysis/conta: add gt file filtering to swap
Summary:
Recent changes introduced to conta require that invalid rsids
and snps with loss of heterozygosity be filtered from the
provided gt files used for swap detection. In addition, filtering
on maf and depth are necessary for better separation of swaps.

Ref T33238, T19074

Test Plan: Run swap and check that the proper snps are filtered

Reviewers: cyakym, osakarya

Reviewed By: osakarya

Subscribers: escott

Maniphest Tasks: T33238, T19074

Differential Revision: https://phabricator.grailbio.com/D42545

fbshipit-source-id: 932b909